fix exchanges, add refresh

+++ examples/thesocialnetwork/ios/TheSocialNetwork/Shared/Views/AccountView.swift
@@ -19,7 +19,7 @@ struct AccountView: View {
         self.user = user
         
         self.formatter = DateFormatter()
-        self.formatter.timeStyle = .short
+        self.formatter.timeStyle = .long
         self.formatter.dateStyle = .short
     }
